pyproject.toml
[project]
name = "codeguard"
version = "0.1.0"
description = "AI Coding Rules for Security and Best Practices"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.2",
]
